feat: Add Ground type for result storage and optimize Patchwork API by removing unnecessary mut
