+++ src/lib.rs
@@ -1,17 +1,19 @@
 mod cloud;
+mod ground;
 mod matrix;
 mod params;
-mod patchworkpp;
+mod patchwork;
 mod points;
 mod utils;
 
 pub use cloud::IntoCloud;
+pub use ground::Ground;
 pub use matrix::{
     matrix_from_raw_buffer, matrix_from_raw_buffer_drop_last, nx3f32_from_raw_buffer,
     nx4f32_from_raw_buffer,
 };
 pub use params::Params;
-pub use patchworkpp::PatchWorkpp;
+pub use patchwork::PatchWork;
 pub use points::{Point3D, Points, Ring, Zone};
 
 mod tests;
